Refactoring - move op::Constant implementation to .cpp file
